refactor(macos): move custom class to individual files
