+++ src/wkwebview/class/mod.rs
@@ -1,0 +1,12 @@
+// Copyright 2020-2024 Tauri Programme within The Commons Conservancy
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: MIT
+
+pub mod wry_web_view;
+pub mod wry_web_view_delegate;
+pub mod document_title_changed_observer;
+pub mod wry_navigation_delegate;
+pub mod wry_download_delegate;
+pub mod wry_web_view_ui_delegate;
+pub mod wry_web_view_parent;
+pub mod url_scheme_handler;
