+++ src/wkwebview/class/wry_download_delegate.rs
@@ -1,0 +1,102 @@
+// Copyright 2020-2024 Tauri Programme within The Commons Conservancy
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: MIT
+
+use std::{path::PathBuf, ptr::null_mut, rc::Rc};
+
+use objc2::{
+  declare_class, msg_send_id, mutability::MainThreadOnly, rc::Retained, runtime::NSObject,
+  ClassType, DeclaredClass,
+};
+use objc2_foundation::{
+  MainThreadMarker, NSData, NSError, NSObjectProtocol, NSString, NSURLResponse, NSURL,
+};
+use objc2_web_kit::{WKDownload, WKDownloadDelegate};
+
+use crate::wkwebview::download::{download_did_fail, download_did_finish, download_policy};
+
+pub struct WryDownloadDelegateIvars {
+  pub started: *mut Box<dyn FnMut(String, &mut PathBuf) -> bool>,
+  pub completed: *mut Rc<dyn Fn(String, Option<PathBuf>, bool)>,
+}
+
+declare_class!(
+  pub struct WryDownloadDelegate;
+
+  unsafe impl ClassType for WryDownloadDelegate {
+    type Super = NSObject;
+    type Mutability = MainThreadOnly;
+    const NAME: &'static str = "WryDownloadDelegate";
+  }
+
+  impl DeclaredClass for WryDownloadDelegate {
+    type Ivars = WryDownloadDelegateIvars;
+  }
+
+  unsafe impl NSObjectProtocol for WryDownloadDelegate {}
+
+  unsafe impl WKDownloadDelegate for WryDownloadDelegate {
+    #[method(download:decideDestinationUsingResponse:suggestedFilename:completionHandler:)]
+    fn download_policy(
+      &self,
+      download: &WKDownload,
+      response: &NSURLResponse,
+      suggested_path: &NSString,
+      handler: &block2::Block<dyn Fn(*const NSURL)>,
+    ) {
+      download_policy(self, download, response, suggested_path, handler);
+    }
+
+    #[method(downloadDidFinish:)]
+    fn download_did_finish(&self, download: &WKDownload) {
+      download_did_finish(self, download);
+    }
+
+    #[method(download:didFailWithError:resumeData:)]
+    fn download_did_fail(
+      &self,
+      download: &WKDownload,
+      error: &NSError,
+      resume_data: &NSData,
+    ) {
+      download_did_fail(self, download, error, resume_data);
+    }
+  }
+);
+
+impl WryDownloadDelegate {
+  pub fn new(
+    download_started_handler: Option<Box<dyn FnMut(String, &mut PathBuf) -> bool>>,
+    download_completed_handler: Option<Rc<dyn Fn(String, Option<PathBuf>, bool)>>,
+    mtm: MainThreadMarker,
+  ) -> Retained<Self> {
+    let started = match download_started_handler {
+      Some(handler) => Box::into_raw(Box::new(handler)),
+      None => null_mut(),
+    };
+    let completed = match download_completed_handler {
+      Some(handler) => Box::into_raw(Box::new(handler)),
+      None => null_mut(),
+    };
+    let delegate = mtm
+      .alloc::<WryDownloadDelegate>()
+      .set_ivars(WryDownloadDelegateIvars { started, completed });
+
+    unsafe { msg_send_id![super(delegate), init] }
+  }
+}
+
+impl Drop for WryDownloadDelegate {
+  fn drop(&mut self) {
+    if self.ivars().started != null_mut() {
+      unsafe {
+        drop(Box::from_raw(self.ivars().started));
+      }
+    }
+    if self.ivars().completed != null_mut() {
+      unsafe {
+        drop(Box::from_raw(self.ivars().completed));
+      }
+    }
+  }
+}
