+++ src/wkwebview/class/wry_web_view_parent.rs
@@ -1,0 +1,50 @@
+// Copyright 2020-2024 Tauri Programme within The Commons Conservancy
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: MIT
+
+use objc2::{
+  declare_class, msg_send_id, mutability::MainThreadOnly, rc::Retained, ClassType, DeclaredClass,
+};
+use objc2_app_kit::{NSApp, NSEvent, NSView};
+use objc2_foundation::MainThreadMarker;
+
+pub struct WryWebViewParentIvars {}
+
+declare_class!(
+  pub struct WryWebViewParent;
+
+  unsafe impl ClassType for WryWebViewParent {
+    type Super = NSView;
+    type Mutability = MainThreadOnly;
+    const NAME: &'static str = "WryWebViewParent";
+  }
+
+  impl DeclaredClass for WryWebViewParent {
+    type Ivars = WryWebViewParentIvars;
+  }
+
+  unsafe impl WryWebViewParent {
+    #[method(keyDown:)]
+    fn key_down(
+      &self,
+      event: &NSEvent,
+    ) {
+      let mtm = MainThreadMarker::new().unwrap();
+      let app = NSApp(mtm);
+      unsafe {
+        if let Some(menu) = app.mainMenu() {
+          menu.performKeyEquivalent(event);
+        }
+      }
+    }
+  }
+);
+
+impl WryWebViewParent {
+  pub fn new(mtm: MainThreadMarker) -> Retained<Self> {
+    let delegate = mtm
+      .alloc::<WryWebViewParent>()
+      .set_ivars(WryWebViewParentIvars {});
+    unsafe { msg_send_id![super(delegate), init] }
+  }
+}
