+++ src/wkwebview/mod.rs
@@ -12,55 +12,46 @@ mod proxy;
 mod synthetic_mouse_events;
 mod util;
 
-use block2::Block;
+mod class;
+use class::url_scheme_handler;
+use class::wry_download_delegate::WryDownloadDelegate;
+pub use class::wry_web_view::WryWebView;
+use class::wry_web_view::WryWebViewIvars;
+use class::wry_web_view_delegate::{WryWebViewDelegate, IPC_MESSAGE_HANDLER_NAME};
+use class::wry_web_view_parent::WryWebViewParent;
+use class::wry_web_view_ui_delegate::WryWebViewUIDelegate;
+use class::{document_title_changed_observer::*, wry_navigation_delegate::WryNavigationDelegate};
 
-use download::{navigation_download_action, navigation_download_response};
 use dpi::{LogicalPosition, LogicalSize};
-use navigation::{
-  did_commit_navigation, did_finish_navigation, navigation_policy, navigation_policy_response,
-};
 use objc2::{
-  declare::ClassBuilder,
-  declare_class, msg_send, msg_send_id,
-  mutability::{InteriorMutable, MainThreadOnly},
   rc::Retained,
-  runtime::{AnyClass, AnyObject, Bool, MessageReceiver, NSObject, ProtocolObject},
+  runtime::{AnyObject, Bool, NSObject, ProtocolObject},
   ClassType, DeclaredClass,
 };
 use objc2_app_kit::{
-  NSApp, NSApplication, NSAutoresizingMaskOptions, NSEvent, NSModalResponse, NSModalResponseOK,
-  NSOpenPanel, NSTitlebarSeparatorStyle, NSView, NSWindow,
+  NSApplication, NSAutoresizingMaskOptions, NSTitlebarSeparatorStyle, NSView, NSWindow,
 };
 use objc2_foundation::{
-  ns_string, CGPoint, CGRect, CGSize, MainThreadMarker, NSArray, NSBundle, NSData, NSDate,
-  NSDictionary, NSError, NSHTTPURLResponse, NSJSONSerialization, NSKeyValueChangeKey,
-  NSKeyValueObservingOptions, NSMutableDictionary, NSMutableURLRequest, NSNumber,
-  NSObjectNSKeyValueCoding, NSObjectNSKeyValueObserverRegistration, NSObjectProtocol, NSString,
-  NSURLResponse, NSUTF8StringEncoding, NSURL, NSUUID,
+  ns_string, CGPoint, CGRect, CGSize, MainThreadMarker, NSBundle, NSDate, NSError,
+  NSJSONSerialization, NSMutableURLRequest, NSNumber, NSObjectNSKeyValueCoding, NSObjectProtocol,
+  NSString, NSUTF8StringEncoding, NSURL, NSUUID,
 };
 #[cfg(target_os = "ios")]
 use objc2_ui_kit::UIScrollView;
 use objc2_web_kit::{
-  WKAudiovisualMediaTypes, WKDownload, WKDownloadDelegate, WKFrameInfo, WKMediaCaptureType,
-  WKNavigation, WKNavigationAction, WKNavigationActionPolicy, WKNavigationDelegate,
-  WKNavigationResponse, WKNavigationResponsePolicy, WKOpenPanelParameters, WKPermissionDecision,
-  WKScriptMessage, WKScriptMessageHandler, WKSecurityOrigin, WKUIDelegate, WKURLSchemeHandler,
-  WKURLSchemeTask, WKUserContentController, WKUserScript, WKUserScriptInjectionTime, WKWebView,
-  WKWebViewConfiguration, WKWebsiteDataStore,
+  WKAudiovisualMediaTypes, WKURLSchemeHandler, WKUserContentController, WKUserScript,
+  WKUserScriptInjectionTime, WKWebView, WKWebViewConfiguration, WKWebsiteDataStore,
 };
 use once_cell::sync::Lazy;
 use raw_window_handle::{HasWindowHandle, RawWindowHandle};
 
 use std::{
-  borrow::Cow,
   collections::{HashMap, HashSet},
-  ffi::{c_void, CStr},
+  ffi::c_void,
   os::raw::c_char,
   panic::{catch_unwind, AssertUnwindSafe},
-  path::PathBuf,
-  ptr::{null_mut, NonNull},
-  rc::Rc,
-  slice, str,
+  ptr::null_mut,
+  str,
   sync::{Arc, Mutex},
 };
 
@@ -72,22 +63,11 @@ use crate::{
   },
 };
 
-use crate::{
-  wkwebview::download::{download_did_fail, download_did_finish, download_policy},
-  DragDropEvent, Error, PageLoadEvent, Rect, RequestAsyncResponder, Result, WebContext,
-  WebViewAttributes, RGBA,
-};
+use crate::{Error, Rect, RequestAsyncResponder, Result, WebContext, WebViewAttributes, RGBA};
 
-use http::{
-  header::{CONTENT_LENGTH, CONTENT_TYPE},
-  status::StatusCode,
-  version::Version,
-  Request, Response as HttpResponse,
-};
+use http::Request;
 
 use self::util::Counter;
-
-const IPC_MESSAGE_HANDLER_NAME: &str = "ipc";
 
 static COUNTER: Counter = Counter::new();
 static WEBVIEW_IDS: Lazy<Mutex<HashSet<u32>>> = Lazy::new(Default::default);
@@ -189,242 +169,6 @@ impl InnerWebView {
   ) -> Result<Self> {
     let mtm = MainThreadMarker::new().ok_or(Error::NotMainThread)?;
 
-    // Task handler for custom protocol
-    extern "C" fn start_task<'a>(
-      this: &AnyObject,
-      _sel: objc2::runtime::Sel,
-      webview: *mut WryWebView,
-      task: *mut ProtocolObject<dyn WKURLSchemeTask>,
-    ) {
-      unsafe {
-        #[cfg(feature = "tracing")]
-        let span = tracing::info_span!(parent: None, "wry::custom_protocol::handle", uri = tracing::field::Empty)
-          .entered();
-
-        let task_key = (*task).hash(); // hash by task object address
-        let task_uuid = (*webview).add_custom_task_key(task_key);
-
-        let ivar = this.class().instance_variable("webview_id").unwrap();
-        let webview_id: u32 = ivar.load::<u32>(this).clone();
-        let ivar = this.class().instance_variable("function").unwrap();
-        let function: &*mut c_void = ivar.load(this);
-        if !function.is_null() {
-          let function =
-            &mut *(*function as *mut Box<dyn Fn(Request<Vec<u8>>, RequestAsyncResponder)>);
-
-          // Get url request
-          let request = (*task).request();
-          let url = request.URL().unwrap();
-
-          let uri = url.absoluteString().unwrap().to_string();
-
-          #[cfg(feature = "tracing")]
-          span.record("uri", uri.clone());
-
-          // Get request method (GET, POST, PUT etc...)
-          let method = request.HTTPMethod().unwrap().to_string();
-
-          // Prepare our HttpRequest
-          let mut http_request = Request::builder().uri(uri).method(method.as_str());
-
-          // Get body
-          let mut sent_form_body = Vec::new();
-          let body = request.HTTPBody();
-          let body_stream = request.HTTPBodyStream();
-          if let Some(body) = body {
-            let length = body.length();
-            let data_bytes = body.bytes();
-            sent_form_body = slice::from_raw_parts(data_bytes.as_ptr(), length).to_vec();
-          } else if let Some(body_stream) = body_stream {
-            body_stream.open();
-
-            while body_stream.hasBytesAvailable() {
-              sent_form_body.reserve(128);
-              let p = sent_form_body.as_mut_ptr().add(sent_form_body.len());
-              let read_length = sent_form_body.capacity() - sent_form_body.len();
-              let count = body_stream.read_maxLength(NonNull::new(p).unwrap(), read_length);
-              sent_form_body.set_len(sent_form_body.len() + count as usize);
-            }
-
-            body_stream.close();
-          }
-
-          // Extract all headers fields
-          let all_headers = request.allHTTPHeaderFields();
-
-          // get all our headers values and inject them in our request
-          if let Some(all_headers) = all_headers {
-            for current_header in all_headers.allKeys().to_vec() {
-              let header_value = all_headers.valueForKey(current_header).unwrap();
-
-              // inject the header into the request
-              http_request =
-                http_request.header(current_header.to_string(), header_value.to_string());
-            }
-          }
-
-          let respond_with_404 = || {
-            let urlresponse = NSHTTPURLResponse::alloc();
-            let response = NSHTTPURLResponse::initWithURL_statusCode_HTTPVersion_headerFields(
-              urlresponse,
-              &url,
-              StatusCode::NOT_FOUND.as_u16().try_into().unwrap(),
-              Some(&NSString::from_str(
-                format!("{:#?}", Version::HTTP_11).as_str(),
-              )),
-              None,
-            )
-            .unwrap();
-            (*task).didReceiveResponse(&response);
-            // Finish
-            (*task).didFinish();
-          };
-
-          // send response
-          match http_request.body(sent_form_body) {
-            Ok(final_request) => {
-              let responder: Box<dyn FnOnce(HttpResponse<Cow<'static, [u8]>>)> =
-                Box::new(move |sent_response| {
-                  fn check_webview_id_valid(webview_id: u32) -> crate::Result<()> {
-                    if !WEBVIEW_IDS.lock().unwrap().contains(&webview_id) {
-                      return Err(crate::Error::CustomProtocolTaskInvalid);
-                    }
-                    Ok(())
-                  }
-                  /// Task may not live longer than async custom protocol handler.
-                  ///
-                  /// There are roughly 2 ways to cause segfault:
-                  /// 1. Task has stopped. pointer of the task not valid anymore.
-                  /// 2. Task had stopped, but the pointer of the task has allocated to a new task.
-                  ///    Outdated custom handler may call to the new task instance and cause segfault.
-                  fn check_task_is_valid(
-                    webview: &WryWebView,
-                    task_key: usize,
-                    current_uuid: Retained<NSUUID>,
-                  ) -> crate::Result<()> {
-                    let latest_task_uuid = webview.get_custom_task_uuid(task_key);
-                    if let Some(latest_uuid) = latest_task_uuid {
-                      if latest_uuid != current_uuid {
-                        return Err(crate::Error::CustomProtocolTaskInvalid);
-                      }
-                    } else {
-                      return Err(crate::Error::CustomProtocolTaskInvalid);
-                    }
-                    Ok(())
-                  }
-
-                  // FIXME: This is 10000% unsafe. `task` and `webview` are not guaranteed to be valid.
-                  // We should consider use sync command only.
-                  unsafe fn response(
-                    task: *mut ProtocolObject<dyn WKURLSchemeTask>,
-                    webview: *mut WryWebView,
-                    task_key: usize,
-                    task_uuid: Retained<NSUUID>,
-                    webview_id: u32,
-                    url: Retained<NSURL>,
-                    sent_response: HttpResponse<Cow<'_, [u8]>>,
-                  ) -> crate::Result<()> {
-                    check_task_is_valid(&*webview, task_key, task_uuid.clone())?;
-
-                    let content = sent_response.body();
-                    // default: application/octet-stream, but should be provided by the client
-                    let wanted_mime = sent_response.headers().get(CONTENT_TYPE);
-                    // default to 200
-                    let wanted_status_code = sent_response.status().as_u16() as i32;
-                    // default to HTTP/1.1
-                    let wanted_version = format!("{:#?}", sent_response.version());
-
-                    let mut headers = NSMutableDictionary::new();
-
-                    if let Some(mime) = wanted_mime {
-                      headers.insert_id(
-                        NSString::from_str(mime.to_str().unwrap()).as_ref(),
-                        NSString::from_str(CONTENT_TYPE.as_str()),
-                      );
-                    }
-                    headers.insert_id(
-                      NSString::from_str(&content.len().to_string()).as_ref(),
-                      NSString::from_str(CONTENT_LENGTH.as_str()),
-                    );
-
-                    // add headers
-                    for (name, value) in sent_response.headers().iter() {
-                      if let Ok(value) = value.to_str() {
-                        headers.insert_id(
-                          NSString::from_str(name.as_str()).as_ref(),
-                          NSString::from_str(value),
-                        );
-                      }
-                    }
-
-                    let urlresponse = NSHTTPURLResponse::alloc();
-                    let response =
-                      NSHTTPURLResponse::initWithURL_statusCode_HTTPVersion_headerFields(
-                        urlresponse,
-                        &url,
-                        wanted_status_code.try_into().unwrap(),
-                        Some(&NSString::from_str(&wanted_version)),
-                        Some(&headers),
-                      )
-                      .unwrap();
-
-                    check_webview_id_valid(webview_id)?;
-                    check_task_is_valid(&*webview, task_key, task_uuid.clone())?;
-                    (*task).didReceiveResponse(&response);
-
-                    // Send data
-                    let bytes = content.as_ptr() as *mut c_void;
-                    let data = NSData::alloc();
-                    // MIGRATE NOTE: we copied the content to the NSData because content will be freed
-                    // when out of scope but NSData will also free the content when it's done and cause doube free.
-                    let data = NSData::initWithBytes_length(data, bytes, content.len());
-                    check_webview_id_valid(webview_id)?;
-                    check_task_is_valid(&*webview, task_key, task_uuid.clone())?;
-                    (*task).didReceiveData(&data);
-
-                    // Finish
-                    check_webview_id_valid(webview_id)?;
-                    check_task_is_valid(&*webview, task_key, task_uuid.clone())?;
-                    (*task).didFinish();
-
-                    (*webview).remove_custom_task_key(task_key);
-                    Ok(())
-                  }
-
-                  let _ = response(
-                    task,
-                    webview,
-                    task_key,
-                    task_uuid,
-                    webview_id,
-                    url.clone(),
-                    sent_response,
-                  );
-                });
-
-              #[cfg(feature = "tracing")]
-              let _span = tracing::info_span!("wry::custom_protocol::call_handler").entered();
-              function(final_request, RequestAsyncResponder { responder });
-            }
-            Err(_) => respond_with_404(),
-          };
-        } else {
-          #[cfg(feature = "tracing")]
-          tracing::warn!(
-            "Either WebView or WebContext instance is dropped! This handler shouldn't be called."
-          );
-        }
-      }
-    }
-    extern "C" fn stop_task(
-      _this: &ProtocolObject<dyn WKURLSchemeHandler>,
-      _sel: objc2::runtime::Sel,
-      webview: &mut WryWebView,
-      task: &ProtocolObject<dyn WKURLSchemeTask>,
-    ) {
-      webview.remove_custom_task_key(task.hash());
-    }
-
     let mut wv_ids = WEBVIEW_IDS.lock().unwrap();
     let webview_id = COUNTER.next();
     wv_ids.insert(webview_id);
@@ -432,19 +176,15 @@ impl InnerWebView {
 
     // Safety: objc runtime calls are unsafe
     unsafe {
-      // Config and custom protocol
       let config = WKWebViewConfiguration::new();
-      let mut protocol_ptrs = Vec::new();
 
+      // Incognito mode
       let os_version = util::operating_system_version();
-
       #[cfg(target_os = "macos")]
       let custom_data_store_available = os_version.0 >= 14;
-
       #[cfg(target_os = "ios")]
       let custom_data_store_available = os_version.0 >= 17;
 
-      // Incognito mode
       let data_store = match (
         attributes.incognito,
         custom_data_store_available,
@@ -460,26 +200,11 @@ impl InnerWebView {
         _ => WKWebsiteDataStore::defaultDataStore(),
       };
 
+      // Register Custom Protocols
+      let mut protocol_ptrs = Vec::new();
       for (name, function) in attributes.custom_protocols {
-        let scheme_name = format!("{}URLSchemeHandler", name);
-        let cls = ClassBuilder::new(&scheme_name, NSObject::class());
-        let cls = match cls {
-          Some(mut cls) => {
-            cls.add_ivar::<*mut c_void>("function");
-            cls.add_ivar::<u32>("webview_id");
-            cls.add_method(
-              objc2::sel!(webView:startURLSchemeTask:),
-              start_task as extern "C" fn(_, _, _, _),
-            );
-            cls.add_method(
-              objc2::sel!(webView:stopURLSchemeTask:),
-              stop_task as extern "C" fn(_, _, _, _),
-            );
-            cls.register()
-          }
-          None => AnyClass::get(&scheme_name).expect("Failed to get the class definition"),
-        };
-        let handler: *mut AnyObject = objc2::msg_send![cls, new];
+        let url_scheme_handler_cls = url_scheme_handler::create(&name);
+        let handler: *mut AnyObject = objc2::msg_send![url_scheme_handler_cls, new];
         let function = Box::into_raw(Box::new(function));
         protocol_ptrs.push(function);
 
@@ -1133,631 +858,4 @@ impl Drop for InnerWebView {
 unsafe fn window_position(view: &NSView, x: i32, y: i32, height: f64) -> CGPoint {
   let frame: CGRect = view.frame();
   CGPoint::new(x as f64, frame.size.height - y as f64 - height)
-}
-
-pub struct WryWebViewIvars {
-  is_child: bool,
-  #[cfg(target_os = "macos")]
-  drag_drop_handler: Box<dyn Fn(DragDropEvent) -> bool>,
-  #[cfg(target_os = "macos")]
-  accept_first_mouse: objc2::runtime::Bool,
-  custom_protocol_task_ids: HashMap<usize, Retained<NSUUID>>,
-}
-
-declare_class!(
-  pub struct WryWebView;
-
-  unsafe impl ClassType for WryWebView {
-    type Super = WKWebView;
-    type Mutability = MainThreadOnly;
-    const NAME: &'static str = "WryWebView";
-  }
-
-  impl DeclaredClass for WryWebView {
-    type Ivars = WryWebViewIvars;
-  }
-
-  unsafe impl WryWebView {
-    #[method(performKeyEquivalent:)]
-    fn perform_key_equivalent(
-      &self,
-      _event: &NSEvent,
-    ) -> Bool {
-      // This is a temporary workaround for https://github.com/tauri-apps/tauri/issues/9426
-      // FIXME: When the webview is a child webview, performKeyEquivalent always return YES
-      // and stop propagating the event to the window, hence the menu shortcut won't be
-      // triggered. However, overriding this method also means the cmd+key event won't be
-      // handled in webview, which means the key cannot be listened by JavaScript.
-      if self.ivars().is_child {
-        Bool::NO
-      } else {
-        unsafe {
-          let result: Bool = self.send_super_message(
-            WKWebView::class(),
-            objc2::sel!(performKeyEquivalent:),
-            (_event,),
-          );
-          result
-        }
-      }
-    }
-
-    #[method(acceptsFirstMouse:)]
-    fn accept_first_mouse(
-      &self,
-      _event: &NSEvent,
-    ) -> Bool {
-        self.ivars().accept_first_mouse
-    }
-  }
-
-  // Drag & Drop
-  #[cfg(target_os = "macos")]
-  unsafe impl WryWebView {
-    #[method(draggingEntered:)]
-    fn dragging_entered(
-      &self,
-      drag_info: &ProtocolObject<dyn objc2_app_kit::NSDraggingInfo>,
-    ) -> objc2_app_kit::NSDragOperation {
-      drag_drop::dragging_entered(self, drag_info)
-    }
-
-    #[method(draggingUpdated:)]
-    fn dragging_updated(
-      &self,
-      drag_info: &ProtocolObject<dyn objc2_app_kit::NSDraggingInfo>,
-    ) -> objc2_app_kit::NSDragOperation {
-      drag_drop::dragging_updated(self, drag_info)
-    }
-
-    #[method(performDragOperation:)]
-    fn perform_drag_operation(
-      &self,
-      drag_info: &ProtocolObject<dyn objc2_app_kit::NSDraggingInfo>,
-    ) -> Bool {
-      drag_drop::perform_drag_operation(self, drag_info)
-    }
-
-    #[method(draggingExited:)]
-    fn dragging_exited(
-      &self,
-      drag_info: &ProtocolObject<dyn objc2_app_kit::NSDraggingInfo>,
-    ) {
-      drag_drop::dragging_exited(self, drag_info)
-    }
-  }
-
-  // Synthetic mouse events
-  #[cfg(target_os = "macos")]
-  unsafe impl WryWebView {
-    #[method(otherMouseDown:)]
-    fn other_mouse_down(
-      &self,
-      event: &NSEvent,
-    ) {
-      synthetic_mouse_events::other_mouse_down(self, event)
-    }
-
-    #[method(otherMouseUp:)]
-    fn other_mouse_up(
-      &self,
-      event: &NSEvent,
-    ) {
-      synthetic_mouse_events::other_mouse_up(self, event)
-    }
-  }
-);
-
-// Custom Protocol Task Checker
-impl WryWebView {
-  fn add_custom_task_key(&mut self, task_id: usize) -> Retained<NSUUID> {
-    let task_uuid = NSUUID::new();
-    self
-      .ivars_mut()
-      .custom_protocol_task_ids
-      .insert(task_id, task_uuid.clone());
-    task_uuid
-  }
-  fn remove_custom_task_key(&mut self, task_id: usize) {
-    self.ivars_mut().custom_protocol_task_ids.remove(&task_id);
-  }
-  fn get_custom_task_uuid(&self, task_id: usize) -> Option<Retained<NSUUID>> {
-    self.ivars().custom_protocol_task_ids.get(&task_id).cloned()
-  }
-}
-
-struct WryWebViewDelegateIvars {
-  controller: Retained<WKUserContentController>,
-  ipc_handler: Box<dyn Fn(Request<String>)>,
-}
-
-declare_class!(
-  struct WryWebViewDelegate;
-
-  unsafe impl ClassType for WryWebViewDelegate {
-    type Super = NSObject;
-    type Mutability = MainThreadOnly;
-    const NAME: &'static str = "WryWebViewDelegate";
-  }
-
-  impl DeclaredClass for WryWebViewDelegate {
-    type Ivars = WryWebViewDelegateIvars;
-  }
-
-  unsafe impl NSObjectProtocol for WryWebViewDelegate {}
-
-  unsafe impl WKScriptMessageHandler for WryWebViewDelegate {
-    // Function for ipc handler
-    #[method(userContentController:didReceiveScriptMessage:)]
-    fn did_receive(
-      this: &WryWebViewDelegate,
-      _controller: &WKUserContentController,
-      msg: &WKScriptMessage,
-    ) {
-      // Safety: objc runtime calls are unsafe
-      unsafe {
-        #[cfg(feature = "tracing")]
-        let _span = tracing::info_span!(parent: None, "wry::ipc::handle").entered();
-
-        let ipc_handler = &this.ivars().ipc_handler;
-        let body = msg.body();
-        let is_string = Retained::cast::<NSObject>(body.clone()).isKindOfClass(NSString::class());
-        if is_string {
-          let body = Retained::cast::<NSString>(body);
-          let js_utf8 = body.UTF8String();
-
-          let frame_info = msg.frameInfo();
-          let request = frame_info.request();
-          let url = request.URL().unwrap();
-          let absolute_url = url.absoluteString().unwrap();
-          let url_utf8 = absolute_url.UTF8String();
-
-          if let (Ok(url), Ok(js)) = (
-            CStr::from_ptr(url_utf8).to_str(),
-            CStr::from_ptr(js_utf8).to_str(),
-          ) {
-            ipc_handler(Request::builder().uri(url).body(js.to_string()).unwrap());
-            return;
-          }
-        }
-
-        #[cfg(feature = "tracing")]
-        tracing::warn!("WebView received invalid IPC call.");
-      }
-    }
-  }
-);
-
-impl WryWebViewDelegate {
-  fn new(
-    controller: Retained<WKUserContentController>,
-    ipc_handler: Box<dyn Fn(Request<String>)>,
-    mtm: MainThreadMarker,
-  ) -> Retained<Self> {
-    let delegate = mtm
-      .alloc::<WryWebViewDelegate>()
-      .set_ivars(WryWebViewDelegateIvars {
-        ipc_handler,
-        controller,
-      });
-
-    let delegate: Retained<Self> = unsafe { msg_send_id![super(delegate), init] };
-
-    let proto_delegate = ProtocolObject::from_ref(delegate.as_ref());
-    unsafe {
-      // this will increate the retain count of the delegate
-      delegate.ivars().controller.addScriptMessageHandler_name(
-        proto_delegate,
-        &NSString::from_str(IPC_MESSAGE_HANDLER_NAME),
-      );
-    }
-
-    delegate
-  }
-}
-
-struct DocumentTitleChangedObserverIvars {
-  object: Retained<WryWebView>,
-  handler: Box<dyn Fn(String)>,
-}
-
-declare_class!(
-  struct DocumentTitleChangedObserver;
-
-  unsafe impl ClassType for DocumentTitleChangedObserver {
-    type Super = NSObject;
-    type Mutability = InteriorMutable;
-    const NAME: &'static str = "DocumentTitleChangedObserver";
-  }
-
-  impl DeclaredClass for DocumentTitleChangedObserver {
-    type Ivars = DocumentTitleChangedObserverIvars;
-  }
-
-  unsafe impl DocumentTitleChangedObserver {
-    #[method(observeValueForKeyPath:ofObject:change:context:)]
-    fn observe_value_for_key_path(
-      &self,
-      key_path: Option<&NSString>,
-      of_object: Option<&AnyObject>,
-      _change: Option<&NSDictionary<NSKeyValueChangeKey, AnyObject>>,
-      _context: *mut c_void,
-    ) {
-      if let (Some(key_path), Some(object)) = (key_path, of_object) {
-        if key_path.to_string() == "title" {
-          unsafe {
-            let handler = &self.ivars().handler;
-            // if !handler.is_null() {
-              let title: *const NSString = msg_send![object, title];
-              handler((*title).to_string());
-            // }
-          }
-        }
-      }
-    }
-  }
-
-  unsafe impl NSObjectProtocol for DocumentTitleChangedObserver {}
-);
-
-impl DocumentTitleChangedObserver {
-  fn new(webview: Retained<WryWebView>, handler: Box<dyn Fn(String)>) -> Retained<Self> {
-    let observer = Self::alloc().set_ivars(DocumentTitleChangedObserverIvars {
-      object: webview,
-      handler,
-    });
-
-    let observer: Retained<Self> = unsafe { msg_send_id![super(observer), init] };
-
-    unsafe {
-      observer
-        .ivars()
-        .object
-        .addObserver_forKeyPath_options_context(
-          &observer,
-          &NSString::from_str("title"),
-          NSKeyValueObservingOptions::NSKeyValueObservingOptionNew,
-          null_mut(),
-        );
-    }
-
-    observer
-  }
-}
-
-impl Drop for DocumentTitleChangedObserver {
-  fn drop(&mut self) {
-    unsafe {
-      self
-        .ivars()
-        .object
-        .removeObserver_forKeyPath(&self, &NSString::from_str("title"));
-    }
-  }
-}
-
-pub struct WryNavigationDelegateIvars {
-  pending_scripts: Arc<Mutex<Option<Vec<String>>>>,
-  has_download_handler: bool,
-  navigation_policy_function: Box<dyn Fn(String, bool) -> bool>,
-  download_delegate: Option<Retained<WryDownloadDelegate>>,
-  on_page_load_handler: Option<Box<dyn Fn(PageLoadEvent)>>,
-}
-
-declare_class!(
-  pub struct WryNavigationDelegate;
-
-  unsafe impl ClassType for WryNavigationDelegate {
-    type Super = NSObject;
-    type Mutability = MainThreadOnly;
-    const NAME: &'static str = "WryNavigationDelegate";
-  }
-
-  impl DeclaredClass for WryNavigationDelegate {
-    type Ivars = WryNavigationDelegateIvars;
-  }
-
-  unsafe impl NSObjectProtocol for WryNavigationDelegate {}
-
-  unsafe impl WKNavigationDelegate for WryNavigationDelegate {
-    #[method(webView:decidePolicyForNavigationAction:decisionHandler:)]
-    fn navigation_policy(
-      &self,
-      webview: &WKWebView,
-      action: &WKNavigationAction,
-      handler: &block2::Block<dyn Fn(WKNavigationActionPolicy)>,
-    ) {
-      navigation_policy(self, webview, action, handler);
-    }
-
-    #[method(webView:decidePolicyForNavigationResponse:decisionHandler:)]
-    fn navigation_policy_response(
-      &self,
-      webview: &WKWebView,
-      response: &WKNavigationResponse,
-      handler: &block2::Block<dyn Fn(WKNavigationResponsePolicy)>,
-    ) {
-      navigation_policy_response(self, webview, response, handler);
-    }
-
-    #[method(webView:didFinishNavigation:)]
-    fn did_finish_navigation(
-      &self,
-      webview: &WKWebView,
-      navigation: &WKNavigation,
-    ) {
-      did_finish_navigation(self, webview, navigation);
-    }
-
-    #[method(webView:didCommitNavigation:)]
-    fn did_commit_navigation(
-      &self,
-      webview: &WKWebView,
-      navigation: &WKNavigation,
-    ) {
-      did_commit_navigation(self, webview, navigation);
-    }
-
-    #[method(webView:navigationAction:didBecomeDownload:)]
-    fn navigation_download_action(
-      &self,
-      webview: &WKWebView,
-      action: &WKNavigationAction,
-      download: &WKDownload,
-    ) {
-      navigation_download_action(self, webview, action, download);
-    }
-
-    #[method(webView:navigationResponse:didBecomeDownload:)]
-    fn navigation_download_response(
-      &self,
-      webview: &WKWebView,
-      response: &WKNavigationResponse,
-      download: &WKDownload,
-    ) {
-      navigation_download_response(self, webview, response, download);
-    }
-  }
-);
-
-impl WryNavigationDelegate {
-  fn new(
-    webview: Retained<WryWebView>,
-    pending_scripts: Arc<Mutex<Option<Vec<String>>>>,
-    has_download_handler: bool,
-    navigation_handler: Option<Box<dyn Fn(String) -> bool>>,
-    new_window_req_handler: Option<Box<dyn Fn(String) -> bool>>,
-    download_delegate: Option<Retained<WryDownloadDelegate>>,
-    on_page_load_handler: Option<Box<dyn Fn(PageLoadEvent, String)>>,
-    mtm: MainThreadMarker,
-  ) -> Retained<Self> {
-    let navigation_policy_function = Box::new(move |url: String, is_main_frame: bool| -> bool {
-      if is_main_frame {
-        navigation_handler
-          .as_ref()
-          .map_or(true, |navigation_handler| (navigation_handler)(url))
-      } else {
-        new_window_req_handler
-          .as_ref()
-          .map_or(true, |new_window_req_handler| (new_window_req_handler)(url))
-      }
-    });
-
-    let on_page_load_handler = if let Some(handler) = on_page_load_handler {
-      let custom_handler = Box::new(move |event| {
-        handler(event, url_from_webview(&webview).unwrap_or_default());
-      }) as Box<dyn Fn(PageLoadEvent)>;
-      Some(custom_handler)
-    } else {
-      None
-    };
-
-    let delegate = mtm
-      .alloc::<WryNavigationDelegate>()
-      .set_ivars(WryNavigationDelegateIvars {
-        pending_scripts,
-        navigation_policy_function,
-        has_download_handler,
-        download_delegate,
-        on_page_load_handler,
-      });
-
-    unsafe { msg_send_id![super(delegate), init] }
-  }
-}
-
-pub struct WryDownloadDelegateIvars {
-  started: *mut Box<dyn FnMut(String, &mut PathBuf) -> bool>,
-  completed: *mut Rc<dyn Fn(String, Option<PathBuf>, bool)>,
-}
-
-declare_class!(
-  pub struct WryDownloadDelegate;
-
-  unsafe impl ClassType for WryDownloadDelegate {
-    type Super = NSObject;
-    type Mutability = MainThreadOnly;
-    const NAME: &'static str = "WryDownloadDelegate";
-  }
-
-  impl DeclaredClass for WryDownloadDelegate {
-    type Ivars = WryDownloadDelegateIvars;
-  }
-
-  unsafe impl NSObjectProtocol for WryDownloadDelegate {}
-
-  unsafe impl WKDownloadDelegate for WryDownloadDelegate {
-    #[method(download:decideDestinationUsingResponse:suggestedFilename:completionHandler:)]
-    fn download_policy(
-      &self,
-      download: &WKDownload,
-      response: &NSURLResponse,
-      suggested_path: &NSString,
-      handler: &block2::Block<dyn Fn(*const NSURL)>,
-    ) {
-      download_policy(self, download, response, suggested_path, handler);
-    }
-
-    #[method(downloadDidFinish:)]
-    fn download_did_finish(&self, download: &WKDownload) {
-      download_did_finish(self, download);
-    }
-
-    #[method(download:didFailWithError:resumeData:)]
-    fn download_did_fail(
-      &self,
-      download: &WKDownload,
-      error: &NSError,
-      resume_data: &NSData,
-    ) {
-      download_did_fail(self, download, error, resume_data);
-    }
-  }
-);
-
-impl WryDownloadDelegate {
-  fn new(
-    download_started_handler: Option<Box<dyn FnMut(String, &mut PathBuf) -> bool>>,
-    download_completed_handler: Option<Rc<dyn Fn(String, Option<PathBuf>, bool)>>,
-    mtm: MainThreadMarker,
-  ) -> Retained<Self> {
-    let started = match download_started_handler {
-      Some(handler) => Box::into_raw(Box::new(handler)),
-      None => null_mut(),
-    };
-    let completed = match download_completed_handler {
-      Some(handler) => Box::into_raw(Box::new(handler)),
-      None => null_mut(),
-    };
-    let delegate = mtm
-      .alloc::<WryDownloadDelegate>()
-      .set_ivars(WryDownloadDelegateIvars { started, completed });
-
-    unsafe { msg_send_id![super(delegate), init] }
-  }
-}
-
-impl Drop for WryDownloadDelegate {
-  fn drop(&mut self) {
-    if self.ivars().started != null_mut() {
-      unsafe {
-        drop(Box::from_raw(self.ivars().started));
-      }
-    }
-    if self.ivars().completed != null_mut() {
-      unsafe {
-        drop(Box::from_raw(self.ivars().completed));
-      }
-    }
-  }
-}
-
-struct WryWebViewUIDelegateIvars {}
-
-declare_class!(
-  struct WryWebViewUIDelegate;
-
-  unsafe impl ClassType for WryWebViewUIDelegate {
-    type Super = NSObject;
-    type Mutability = MainThreadOnly;
-    const NAME: &'static str = "WryWebViewUIDelegate";
-  }
-
-  impl DeclaredClass for WryWebViewUIDelegate {
-    type Ivars = WryWebViewUIDelegateIvars;
-  }
-
-  unsafe impl NSObjectProtocol for WryWebViewUIDelegate {}
-
-  unsafe impl WKUIDelegate for WryWebViewUIDelegate {
-    #[method(webView:runOpenPanelWithParameters:initiatedByFrame:completionHandler:)]
-    fn run_file_upload_panel(
-      &self,
-      _webview: &WryWebView,
-      open_panel_params: &WKOpenPanelParameters,
-      _frame: &WKFrameInfo,
-      handler: &block2::Block<dyn Fn(*const NSArray<NSURL>)>
-    ) {
-      unsafe {
-        if let Some(mtm) = MainThreadMarker::new() {
-          let open_panel = NSOpenPanel::openPanel(mtm);
-          open_panel.setCanChooseFiles(true);
-          let allow_multi = open_panel_params.allowsMultipleSelection();
-          open_panel.setAllowsMultipleSelection(allow_multi);
-          let allow_dir = open_panel_params.allowsDirectories();
-          open_panel.setCanChooseDirectories(allow_dir);
-          let ok: NSModalResponse = open_panel.runModal();
-          if ok == NSModalResponseOK {
-            let url = open_panel.URLs();
-            (*handler).call((Retained::as_ptr(&url),));
-          } else {
-            (*handler).call((null_mut(),));
-          }
-        }
-      }
-    }
-
-    #[method(webView:requestMediaCapturePermissionForOrigin:initiatedByFrame:type:decisionHandler:)]
-    fn request_media_capture_permission(
-      &self,
-      _webview: &WryWebView,
-      _origin: &WKSecurityOrigin,
-      _frame: &WKFrameInfo,
-      _capture_type: WKMediaCaptureType,
-      decision_handler: &Block<dyn Fn(WKPermissionDecision)>
-    ) {
-      //https://developer.apple.com/documentation/webkit/wkpermissiondecision?language=objc
-      (*decision_handler).call((WKPermissionDecision::Grant,));
-    }
-  }
-);
-
-impl WryWebViewUIDelegate {
-  fn new(mtm: MainThreadMarker) -> Retained<Self> {
-    let delegate = mtm
-      .alloc::<WryWebViewUIDelegate>()
-      .set_ivars(WryWebViewUIDelegateIvars {});
-    unsafe { msg_send_id![super(delegate), init] }
-  }
-}
-
-struct WryWebViewParentIvars {}
-
-declare_class!(
-  struct WryWebViewParent;
-
-  unsafe impl ClassType for WryWebViewParent {
-    type Super = NSView;
-    type Mutability = MainThreadOnly;
-    const NAME: &'static str = "WryWebViewParent";
-  }
-
-  impl DeclaredClass for WryWebViewParent {
-    type Ivars = WryWebViewParentIvars;
-  }
-
-  unsafe impl WryWebViewParent {
-    #[method(keyDown:)]
-    fn key_down(
-      &self,
-      event: &NSEvent,
-    ) {
-      let mtm = MainThreadMarker::new().unwrap();
-      let app = NSApp(mtm);
-      unsafe {
-        if let Some(menu) = app.mainMenu() {
-          menu.performKeyEquivalent(event);
-        }
-      }
-    }
-  }
-);
-
-impl WryWebViewParent {
-  fn new(mtm: MainThreadMarker) -> Retained<Self> {
-    let delegate = mtm
-      .alloc::<WryWebViewParent>()
-      .set_ivars(WryWebViewParentIvars {});
-    unsafe { msg_send_id![super(delegate), init] }
-  }
 }
